Moved the contents src/common.h to src/utilities.h, as originally planned.
