+++ src/CMakeLists.txt
@@ -102,7 +102,6 @@ set(MODULE_HEADER_FILES
 )
 
 set(GIT_HEADER_FILES
-  ${CMAKE_CURRENT_SOURCE_DIR}/common.h
   ${CMAKE_CURRENT_SOURCE_DIR}/namespaces.h
   ${CMAKE_CURRENT_SOURCE_DIR}/utilities.h
   ${CMAKE_CURRENT_SOURCE_DIR}/xmlattribute.h
