+++ src/validator.cpp
@@ -14,7 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#include "common.h"
 #include "namespaces.h"
 #include "utilities.h"
 #include "xmldoc.h"
